Extract ShowOpenTablesStatement to reduce duplicate logic for MySQL and doris
